requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
astroid=2.1.0=py36_0
autopep8=1.4.3=py36_0
ca-certificates=2019.1.23=0
certifi=2018.11.29=py36_0
django=2.1.5=py36_0
djangorestframework=3.9.1=py_0
isort=4.3.4=py36_0
lazy-object-proxy=1.3.1=py36h14c3975_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libstdcxx-ng=8.2.0=hdf63c60_1
mccabe=0.6.1=py36_1
mysql-connector-c=6.1.11=h597af5e_0
mysqlclient=1.3.14=py36h7b6447c_0
ncurses=6.1=he6710b0_1
openssl=1.1.1a=h7b6447c_0
pip=19.0.1=py36_0
pycodestyle=2.5.0=py36_0
pylint=2.2.2=py36_0
python=3.6.8=h0371630_0
pytz=2018.9=py36_0
readline=7.0=h7b6447c_5
setuptools=40.7.3=py36_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
typed-ast=1.1.0=py36h14c3975_0
wheel=0.32.3=py36_0
wrapt=1.11.1=py36h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
